CAN driver module converted into a package
